Linux build for deb and AppImage
Additionally improved the way of building applications
by using a node script and using electron-builder in it.

Made a more portable way of building that works on Linux,
Mac, and theoretically on Windows (not tested yet).

Also added several build resources such as icons and others
and removed unneded files.