Avoid Auto Translation
When using Chrome with the automatic translation featured turned on, weird behaviour occurs both in the board and in the code editor. The simple HTML tags added should request translations services to avoid translating automatically. Also the language has been changed to Spanish, as most of our users are until we come up with a real i18n solution.